Moved internet room share to a separate page
