+++ src/main.ts
@@ -39,9 +39,9 @@ const comps = [
   Tabs,
   Table,
   Tag,
+  Toast,
   Tooltip,
   Snackbar,
-  Toast,
   Upload
 ]
 
@@ -65,12 +65,18 @@ if (window.location.hostname === 'localhost') {
   announceURLs = ['ws://0.0.0.0:5000']
 }
 
+const INTERNET_ROOM_CODE_LENGTH = 4
+
 Vue.prototype.$ANNOUNCE_URLS = announceURLs
-Vue.prototype.$INTERNET_ROOM_CODE_LENGTH = 4
+Vue.prototype.$INTERNET_ROOM_CODE_LENGTH = INTERNET_ROOM_CODE_LENGTH
 Vue.prototype.$INTERNET_ROOM_SHARE_LINK = 'https://WebDrop.Space/#/?room='
 
 Vue.prototype.$p2pt = null
 Vue.prototype.$pf = new PeerFiles()
+
+Vue.prototype.$validateRoomCode = (roomCode: string) => {
+  return roomCode.length === INTERNET_ROOM_CODE_LENGTH
+}
 
 store.commit('initSettings')
